Alteração na Tabela Grupo
Estou alterando para pedir confirmação para deletar um Grupo;

+++ grupo/lista_grupo.php
@@ -19,7 +19,7 @@
 	    <td><?php echo $linha['nome']; ?></td>
 	          <td>
                 <a class="btn btn-primary btn-sm" href="grupo.php?acao=buscar&id=<?php echo $linha['id']; ?>">Editar</a>
-                <a class="btn btn-danger btn-sm" href="grupo.php?acao=excluir&id=<?php echo $linha['id']; ?>">Excluir</a>
+                <a class="btn btn-danger btn-sm" onclick="return confirm('Deseja mesmo Deletar o Grupo?')" href="grupo.php?acao=excluir&id=<?php echo $linha['id']; ?>">Excluir</a>
             </td>
           </tr>
         <?php endforeach; ?>
